Improved caching, begin seperating logic.
